# gcp/apigateway/openapi.yaml
swagger: '2.0'
info:
  title: Mrs Santa Preflight API Gateway
  description: A Santa Sync Server
  version: 0.0.1
schemes:
  - https
consumes:
  - application/json 
produces:
  - application/json
paths:
...
  /api/rules/{id}:
    get:
      summary: "Handle api requests"
      operationId: apirule
      x-google-backend:
        address: https://mrssanta-api-3mfoapoj6a-ew.a.run.app/rules/{id}
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: "ID of the rule"
      responses:
        200:
          description: "Successful response"
          schema:
            $ref: '#/definitions/AnyValue'
definitions:
  AnyValue: {}
...
securityDefinitions:
  apiKey:
    type: apiKey
    in: header
    name: X-API-KEY